+++ src/main/java/step3/domain/RandomLottoNumGenerator.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.domain;
 
 import java.util.ArrayList;
 import java.util.Arrays;
