+++ src/main/java/step3/domain/SameLottoChecker.java
@@ -1,4 +1,6 @@
-package step3;
+package step3.domain;
+
+import step3.domain.Lotto;
 
 public class SameLottoChecker {
