+++ src/test/java/step3/LottoNumberTest.java
@@ -1,6 +1,7 @@
 package step3;
 
 import org.junit.jupiter.api.Test;
+import step3.domain.LottoNumber;
 
 import static org.assertj.core.api.Assertions.assertThat;
