+++ src/test/java/step3/MoneyTest.java
@@ -2,6 +2,7 @@ package step3;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step3.domain.Money;
 import step3.exception.NotEnoughInitMoneyException;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
