+++ src/test/java/step3/SameLottoCheckerTest.java
@@ -2,9 +2,10 @@ package step3;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step3.domain.Lotto;
+import step3.domain.SameLottoChecker;
 
 import java.util.Arrays;
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
